requirements.txt
fastapi[standard]
sqlalchemy
psycopg2-binary
passlib
bcrypt
alembic
python-jose
aiofiles
requests
pytest
jinja2
websockets